Factor out sidebar component in wallet app
